(PUP-5082) Use Compiler namespace and validation_stage
This commit changes the following:
1. Moves the catalog validators used by the compiler from the
   Puppet::Parser namespace into Puppet::Parser::Compiler.
2. Merges the CatalogValidator and CatalogValidators namespaces
   into just CatalogValidator.
3. Changes the class method `stage?` into the more descriptive
   `validation_stage?`
4. Breaks some long single line statements into several lines.
