+++ lib/puppet/parser.rb
@@ -3,7 +3,7 @@ module Puppet::Parser; end
 
 require 'puppet/parser/ast'
 require 'puppet/parser/compiler'
-require 'puppet/parser/catalog_validator'
+require 'puppet/parser/compiler/catalog_validator'
 require 'puppet/resource/type_collection'
 
 require 'puppet/parser/functions'
@@ -14,7 +14,7 @@ require 'puppet/resource/type_collection_helper'
 require 'puppet/resource/type'
 require 'monitor'
 
-Dir[File.dirname(__FILE__) + '/parser/catalog_validators/*.rb'].each { |f| require f }
+Dir[File.dirname(__FILE__) + '/parser/compiler/catalog_validator/*.rb'].each { |f| require f }
 
 # PUP-3274 This should probably go someplace else
 class Puppet::LexError < RuntimeError; end
